Small refactor
Moved api test to correct file
Changed visibility of private functions
